# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Film API",
        "description": "Film API Documentation",
        "contact": {
            "name": "Paul Bouvignies",
            "email": "paul.bouvignies@ynov.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Film API Server"
        }
    ],
    "paths": {
        "/categories": {
            "get": {
                "tags": [
                    "Category"
                ],
                "summary": "Get all categories",
                "operationId": "index",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Category"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Category"
                ],
                "summary": "Add a new category",
                "operationId": "store",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Category"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Category"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/categories/{id}": {
            "get": {
                "tags": [
                    "Category"
                ],
                "summary": "Get films by category",
                "operationId": "getCatgory",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Category ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Film"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Category not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Category"
                ],
                "summary": "Update a category",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Category ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Category"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Category"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Category not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/films": {
            "get": {
                "tags": [
                    "Films"
                ],
                "summary": "Get all films",
                "operationId": "6f65b2a58d7d23ba8303a3417286ca1e",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    },
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Search query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "items_per_page",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Film"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Films"
                ],
                "summary": "Create a film",
                "operationId": "054fce5034b2ae9d31c042cf63ad433e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Film"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "422": {
                        "description": "Validation errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Validation errors"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/films/{id}": {
            "get": {
                "tags": [
                    "Films"
                ],
                "summary": "Get a film",
                "operationId": "18e4aba7bc625926b361e99b4c512c87",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the film",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Film"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Film not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Films"
                ],
                "summary": "Update a film",
                "operationId": "5b9486f28697daf734d53f9141bef3b9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the film",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Film"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Film"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Film not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Films"
                ],
                "summary": "Delete a film",
                "operationId": "4659b6faaa14326138d081a273e6c413",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the film",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Film not found"
                    }
                }
            }
        },
        "/films/{id}/categories": {
            "get": {
                "tags": [
                    "Films"
                ],
                "summary": "Get all categories of a film",
                "operationId": "ab3e38ee8f8b0c29b172c784a6abb87f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the film",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Category"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Film not found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Film": {
                "title": "Film",
                "required": [
                    "title"
                ],
                "properties": {
                    "id": {
                        "description": "The film ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "title": {
                        "description": "The film title",
                        "type": "string"
                    },
                    "release_date": {
                        "description": "The film release date",
                        "type": "string",
                        "format": "date"
                    },
                    "runtime": {
                        "description": "The film runtime",
                        "type": "integer"
                    },
                    "director": {
                        "description": "The film director",
                        "type": "string"
                    },
                    "plot": {
                        "description": "The film plot",
                        "type": "string"
                    },
                    "rating": {
                        "description": "The film rating",
                        "type": "number",
                        "format": "integer"
                    },
                    "posterUrl": {
                        "description": "The film poster URL",
                        "type": "string"
                    },
                    "linked_categorie": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "description": "The category ID",
                                    "type": "integer",
                                    "format": "int64"
                                },
                                "name": {
                                    "description": "The category name",
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
            "Category": {
                "title": "Category",
                "required": [
                    "nom"
                ],
                "properties": {
                    "id": {
                        "description": "The category ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "nom": {
                        "description": "The category name",
                        "type": "string"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Category",
            "description": "Category API"
        },
        {
            "name": "Films",
            "description": "API Endpoints of Films"
        }
    ]
}